Daemon: improve read/send loop
With this change the daemon will send serialized input records in
"batches". Instead of sending all available records in a row it will
send N in a row, than wait for 1 millisecond, then send N more (if
available).
While this doesn't fix the copy/paste problem it does already improve
the situation.