Fixed a bug preventing minimization of individual client consoles
Activating the daemon console will cause all client consoles to be moved
to the foreground.
When minimizing a window the focus will change back to the previously
focussed window.
Minimizing a client window with the daemon window selected would thus
prevent the client from being minimized because the daemon window would
get focus right after bringing it back to the foreground.