Fix error handling when reading registry values
If we cannot read the value, i.e. open it, that might be a priviledge
issue or the location might have changed.
This might happen with a future windows update or even already now on
windows 11. We should emit a warning about this but not crash the
program.
If the data contained in the key has not the expected type that is a
major bug and we will crash just to make sure we get notified about it
by users running into the issue