Support missing DelegationConsole/Terminal registry values
If the default terminal application setting was never changed from its
default "Let windows decide", the DelegationConsole and
DelegationTerminal registry values won't exist.
Prior to this change this would have prevented us from retrieving the
original configuration and thus setting it to conhost.exe.
If the registry values are missing we consider the default to be the
original configuration and proceed with setting it to conhost and later
on reverting back to the original configuration.