Fix flicker when using `[c]reate window` control mode option
Especially when creating few windows via this option the initial sizing
of the window would occupy the entire workspace area and only
afterwards get resized based on the number of other open client windows.
With this change the initial size already accounts for the total
number of client windows, meaning no noticable resizing is necesary
anymore.