Fix broken client window rendering after `[c]reate window`
It seems that since windows update `10.0.19041.5072` the rendering of a
client window can break. Either not rendering the content or not
rendering the entire window including borders.
For unknown reasons resizing the affected window helps.
We cannot easily detect which window is affected, so we draw all windows
1 pixel in each dimension too small without repainting them, then
apply the correct size and repaint.
This solve the issue without much overhead.