news-fragment-check: separate check and PR comment
Instead of creating the comment content and posting it in the same
workflow with a `pull_request` trigger, only the comment content is
created on a `pull_request` trigger, the posting is done in a dedicated
workflow using the `workflow_run` trigger.
This allows PRs from forks to still receive comments on their PRs.

GitHub: #133