news-fragment-check: use post-pr-comment
With this change the news-fragment-check workflow merely generates the contents of the comment.
The PR comment itself will be posted by the post-pr-comment workflow.
This separation is necessary as PRs from forks using the `pull_request` trigger do not have the required permissions to post comments on PRs while workflows using the `workflow_run` trigger come from the base branch and have the required permissions.

GitHub: #133